cases/browser_use/requirements.txt
browser-use
langchain-google-genai
python-dotenv